+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Fixed forwardRef description ([#312](https://github.com/Roblox/roact/pull/312)).
+* Silenced indentation warning in script analysis
 
 ## [1.4.0](https://github.com/Roblox/roact/releases/tag/v1.4.0) (November 19th, 2020)
 * Introduce forwardRef ([#307](https://github.com/Roblox/roact/pull/307)).